requirements.txt
about-time==4.2.1
alive-progress==3.1.5
grapheme==0.6.0
numpy==2.0.1
pillow==10.4.0
